基本完成

+++ run.py
@@ -1,13 +1,11 @@
 import taptap
-import resource
-import gameInformation
 
 import requests
 import re
 from urllib.parse import unquote, urlparse
 import os
 
-def download_file_minimal(url, filename=None):
+def download_file_minimal(url, filename=None) -> str:
     """
     最简洁版本，只显示百分比进度
     """
@@ -20,6 +18,8 @@ def download_file_minimal(url, filename=None):
         total_size = int(response.headers.get('content-length', 0))
         downloaded_size = 0
         filename = re.search(r'filename="([^"]+)"',response.headers.get('content-disposition', '')).group(1)
+        if(os.path.exists(filename)): # 本地减少每次下载花费的时间
+            return filename
         with open(filename, 'wb') as file:
             for chunk in response.iter_content(chunk_size=8192):
                 if chunk:
@@ -29,7 +29,7 @@ def download_file_minimal(url, filename=None):
                     if total_size > 0:
                         percent = (downloaded_size / total_size) * 100
                         print(f"\r{percent:.1f}%", end='', flush=True)
-        
+            return filename
     except Exception as e:
         raise e
 try:
@@ -44,6 +44,17 @@ except Exception as e:
     print(e)
     exit(1)
     
-os.system(f"python Resource.py {file_name}")
+version = (file_name.split('-')[1]).split('.')[0]
+try:
+    with open('VERSION','r',encoding='UTF-8')as f:
+        if(int(f.read()) >= int(version)): pass
+        else: 
+            print('fff')
+            raise Exception()
+except:            
+    with open("VERSION","w",encoding='UTF-8') as f:
+        f.write(version)
+        f.close()
+os.system(f"python resource.py {file_name}")
 os.system(f"python gameInformation.py {file_name}")
 os.system("python replaceAvatarName.py")
